✨ refactor(repository): move maize performance repo to its own file
Extract MaizePerformanceRepo from the Dao file into a dedicated class
file to improve code organization and maintainability.
